pyproject.toml
[tool.poetry]
name = "task-scheduler-back"
version = "0.1.0"
description = ""
authors = ["moradalkhatib <muradalkhateb@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
whitenoise = "^5.3.0"
django-cors-headers = "^3.10.1"
pytz = "^2021.3"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
